Pipenv support + minor improvements

+++ python/aliases.zsh
@@ -1,6 +1,3 @@
-alias jupstart="jupyter lab"
-alias jupstartpoetry="poetry run jupyter lab"
-
 # Virtual Environment
 crvenv() {
     local venv_nm=.venv
@@ -18,22 +15,6 @@ crvenv() {
     python -m venv $venv_nm
 }
 
-crpoetry() {
-    local name=project
-    local python_ver=3.7.3
-    while [[ "$#" -gt 0 ]]
-    do
-        case $1 in
-        -v|--version) local python_ver=${2:-3.7.3} ;;
-        -n|--name) local name=${2:-project} ;;
-        esac
-        shift
-    done
-    pyenv install -s $python_ver
-    pyenv local $python_ver
-    poetry new $name
-}
-
 launchvenv() {
     local venv_nm=${1:-.venv}
     source "$venv_nm"/bin/activate
@@ -42,23 +23,37 @@ launchvenv() {
 
 # iPython Kernel
 setkernel() {
-    pip install ipykernel
-    python -m ipykernel install --user --name "$1" --display-name "$2"
-}
-
-poetrykernel() {
-    poetry add -D ipykernel
-    poetry run python -m ipykernel install --user --name "$1"
+    if [ -f pyproject.toml ]; then
+        poetry add --dev ipykernel
+    elif [ -f Pipfile ]; then
+        pipenv install --dev ipykernel
+    else
+        pip install ipykernel
+    fi
+    source $1/bin/activate
+    python -m ipykernel install --user --name $(echo "${PWD##*/}")-venv
+    deactivate
 }
 
 
 # Jupyter
-jupscrpoetry() {
-    echoInfo "Creating screen `cur_dir`-screen"
-    screen -dmS `cur_dir`-screen poetry run jupyter lab
+jupstart() {
+    if [ -f pyproject.toml ]; then
+        poetry run jupyter lab
+    elif [ -f Pipfile ]; then
+        pipenv run jupyter lab
+    else
+        jupyter lab
+    fi
 }
 
 jupscr() {
-    echoInfo "Creating screen `cur_dir`-screen"
-    screen -dmS `cur_dir`-screen jupyter lab
+    echoInfo "Creating screen $(cur_dir)-screen"
+    if [ -f pyproject.toml ]; then
+        screen -dmS $(cur_dir)-screen poetry run jupyter lab
+    elif [ -f Pipfile ]; then
+        screen -dmS $(cur_dir)-screen pipenv run jupyter lab
+    else
+        screen -dmS $(cur_dir)-screen jupyter lab
+    fi
 }
